[refactor] Deliverer- 4계층 수정
